requirements.txt
requests~=2.31.0
pillow~=10.3.0
selenium~=4.20.0
validators~=0.28.1
Flask~=3.0.3
beautifulsoup4~=4.12.3
pytesseract~=0.3.10
redis~=5.0.4
Flask-Caching~=2.3.0
python-dotenv~=1.0.1
